// Web app/csi-mkd-premarital-app-FE/src/api/api-main-app/models/class-feedback-detail-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClassFeedbackRatingsDto } from '../models/class-feedback-ratings-dto';
import { ClassFeedbackTextResponsesDto } from '../models/class-feedback-text-responses-dto';
export interface ClassFeedbackDetailDto {
  date: string;
  instructorId: number;
  ratings: ClassFeedbackRatingsDto;
  textResponses: ClassFeedbackTextResponsesDto;
}